requirements.txt
flask==2.3.3
pandas==2.0.3
openpyxl==3.1.2
numpy==1.25.2
requests==2.31.0
python-dotenv==1.0.0
flask-wtf==1.1.1
xlsxwriter==3.1.0
anthropic==0.5.0